genlearn-ai/backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pandas>=2.2.0
python-dotenv==1.0.0
httpx>=0.27.0
bcrypt==4.1.2
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
pydantic[email]==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6
aiofiles==23.2.1
pillow>=10.3.0
requests>=2.31.0
